pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "wildlife-detection"
version = "0.1.0"
description = "Aerial wildlife detection pipeline for visual images"
requires-python = ">=3.10"

# Base dependencies — all environments need these.
# PyTorch from pip is CPU-only. For GPU/MPS acceleration,
# conda install pytorch torchvision -c pytorch  (before pip install)
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "pillow",
    "pandas",
    "scipy",
    "matplotlib",
    "tqdm",
    "albumentations",
    "huggingface_hub",
    "requests",
    "python-dotenv",
    "loguru",
]

[project.optional-dependencies]

# ── Environment extras (match the three conda envs) ─────────────────
#
# Local:  conda env create -f environment-<name>.yml
# Colab:  !pip install ".[megadetector]"  (pytorch is pre-installed)

# fit-megadetector (P1–P2): data exploration + MegaDetector inference
# Note: megadetector pins protobuf<=3.20.1, conflicts with wandb/label-studio
megadetector = [
    "megadetector",
]

# fit-training (P3–P8): full training & inference pipeline
# Does NOT include megadetector (protobuf conflict) or GDAL
training = [
    "ultralytics",
    "sahi",
    "timm",
    "segment-anything",
    "scikit-learn",
    "wget",
    "tqdm",
    "wandb",
    "speciesnet",
]

# fit-herdnet: geospatial pipeline (extends training with GDAL/rasterio)
# GDAL/rasterio must be installed via conda (not pip).
# Only needed for geospatial workflows (GeoTIFF tiling, coordinate transforms).
herdnet = [
    "wildlife-detection[training]",
    "animaloc @ git+https://github.com/cwinkelmann/HerdNet.git",
]

# ── Standalone extras ───────────────────────────────────────────────

# Label Studio for annotation (P2): needs protobuf>=6.31.1,
# conflicts with megadetector. Best run via Docker instead:
#   docker run -p 8080:8080 humansignal/label-studio
labelstudio = [
    "label-studio",
]

# Development tools (not needed in Colab)
dev = [
    "pytest",
    "jupyterlab",
    "jupytext",
    "nbconvert",
    "ipykernel",

]

animl = [
    "animl"
]

[tool.setuptools.packages.find]
where = ["src"]